extract DynamoDBBatchError
